Extract common classes not tied to tile service
Extracting ModelingTypes and ModelingException helps reduce the size of
ModelingServiceActor and get it closer to the ideal one object/class per
file. The vector handling operations will be useful in both the tile
service and the summary service.
