+++ common/src/main/scala/ModelingTypes.scala
@@ -1,0 +1,7 @@
+package org.opentreemap.modeling
+
+object ModelingTypes {
+  // Map of layer names to selected values.
+  // Ex. { Layer1: [1, 2, 3], ... }
+  type LayerMaskType = Map[String, Array[Int]]
+}
